Refactor code, add more code docs
Add additional documentation to the code for the sake of more easily
enabling others to understand it / grok through it.

Remove dead code and remove remnants from previous rewrites.
