--- internal/dockerhub/package.go
@@ -1,5 +1,0 @@
-// Package docker provides a client using the Docker API. Lookup of the latest
-// image borrows from Renovate.
-// SEE: https://github.com/renovatebot/renovate/tree/main/lib/modules/datasource/docker
-// SEE: https://github.com/renovatebot/renovate/tree/main/lib/modules/versioning/docker
-package dockerhub
