+++ internal/events/hub.go
@@ -5,6 +5,7 @@ import (
 	"sync"
 )
 
+// Hub is an event hub allowing broadcasting of events to subscribers.
 type Hub[T any] struct {
 	mutex    sync.Mutex
 	channels map[chan T]struct{}
@@ -12,10 +13,14 @@ type Hub[T any] struct {
 
 func NewHub[T any]() *Hub[T] {
 	return &Hub[T]{
-		channels: make(map[chan T]struct{}, 0),
+		channels: make(map[chan T]struct{}),
 	}
 }
 
+// Subscribe to events.
+// Returns a channel which will receives all broadcast events.
+// The channel is closed whenever the context expires.
+// Subscribers should receive events from the channel in a timely manner.
 func (h *Hub[T]) Subscribe(ctx context.Context) <-chan T {
 	h.mutex.Lock()
 	defer h.mutex.Unlock()
@@ -37,6 +42,8 @@ func (h *Hub[T]) Subscribe(ctx context.Context) <-chan T {
 	return ch
 }
 
+// Broadcast an event to all subscribers synchronously.
+// Returns an error if the context expires before all subscribers were invoked.
 func (h *Hub[T]) Broadcast(ctx context.Context, event T) error {
 	h.mutex.Lock()
 	defer h.mutex.Unlock()
