+++ internal/gitlab/client.go
@@ -21,6 +21,7 @@ type Client struct {
 	Client *httputil.Client
 }
 
+// GetRepositoryDescription retrieves the description of a repository.
 func (c *Client) GetRepositoryDescription(ctx context.Context, fullPath string) (string, error) {
 	payload, err := json.Marshal(map[string]any{
 		"operationName": "getProject",
@@ -71,6 +72,7 @@ func (c *Client) GetRepositoryDescription(ctx context.Context, fullPath string) 
 	return result.Data.Project.Description, nil
 }
 
+// GetRepositoryREADMEBlob returns a repository's README blob.
 func (c *Client) GetRepositoryREADMEBlob(ctx context.Context, fullPath string) (*Blob, error) {
 	// https://gitlab.com/arm-research/smarter/smarter-device-manager/-/blob/master/README.md?format=json&viewer=rich
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, "https://gitlab.com/"+fullPath, nil)
@@ -105,6 +107,8 @@ func (c *Client) GetRepositoryREADMEBlob(ctx context.Context, fullPath string) (
 	return c.GetBlob(ctx, string(readmeHref[1]), true)
 }
 
+// GetBlob retrieves the blob at href.
+// Set includeRaw to true to download the blob's contents.
 func (c *Client) GetBlob(ctx context.Context, href string, includeRaw bool) (*Blob, error) {
 	u := url.URL{
 		Scheme:   "https",
@@ -166,6 +170,8 @@ func (c *Client) GetBlob(ctx context.Context, href string, includeRaw bool) (*Bl
 	return &blob, nil
 }
 
+// GetRegistryToken returns a token for use with Docker Hub with pull
+// permissions on the specified repository.
 func (c *Client) GetRegistryToken(ctx context.Context, repository string) (string, error) {
 	// TODO: Registries expose the realm and scheme via Www-Authenticate if 403
 	// is given
@@ -205,6 +211,7 @@ func (c *Client) GetRegistryToken(ctx context.Context, repository string) (strin
 	return result.Token, nil
 }
 
+// HandleAuth authenticates a request to the GitLab registry.
 func (c *Client) HandleAuth(r *http.Request) error {
 	name := oci.NameFromAPI(r.URL.Path)
 	// lscr.io is a pseudo-registry that forwards to one of multiple backends,
@@ -221,141 +228,4 @@ func (c *Client) HandleAuth(r *http.Request) error {
 	r.Header.Set("Authorization", "Bearer "+token)
 
 	return nil
-}
-
-func (c *Client) GetProjectContainerRepositories(ctx context.Context, fullPath string) ([]ContainerRepository, error) {
-	payload, err := json.Marshal(map[string]any{
-		"operationName": "getProjectContainerRepositories",
-		"variables": map[string]any{
-			"sort":     "UPDATED_DESC",
-			"fullPath": fullPath,
-			"first":    20,
-		},
-		"query": `query getProjectContainerRepositories($fullPath: ID!, $name: String, $first: Int, $last: Int, $after: String, $before: String, $sort: ContainerRepositorySort) {
-  project(fullPath: $fullPath) {
-    id
-    containerRepositories(
-      name: $name
-      after: $after
-      before: $before
-      first: $first
-      last: $last
-      sort: $sort
-    ) {
-      nodes {
-        id
-        location
-      }
-    }
-  }
-}`,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, "https://gitlab.com/api/graphql", bytes.NewReader(payload))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// TODO: The cache doesn't understand graphql, so we can't cache this request
-	res, err := c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	if err := httputil.AssertStatusCode(res, http.StatusOK); err != nil {
-		return nil, err
-	}
-
-	var result struct {
-		Errors []any `json:"error"`
-		Data   struct {
-			Project struct {
-				ContainerRepositories struct {
-					Nodes []ContainerRepository `json:"nodes"`
-				} `json:"containerRepositories"`
-			} `json:"project"`
-		} `json:"data"`
-	}
-	if err := json.NewDecoder(res.Body).Decode(&result); err != nil {
-		return nil, err
-	} else if len(result.Errors) > 0 {
-		return nil, fmt.Errorf("got unexpected graphql error: %v", result.Errors)
-	}
-
-	return result.Data.Project.ContainerRepositories.Nodes, nil
-}
-
-func (c *Client) GetProjectContainerRepositoryTags(ctx context.Context, id string) ([]ContainerRepositoryTag, error) {
-	payload, err := json.Marshal(map[string]any{
-		"operationName": "getContainerRepositoryTags",
-		"variables": map[string]any{
-			"referrers": true,
-			"id":        id,
-			"first":     20,
-			"sort":      "PUBLISHED_AT_DESC",
-		},
-		"query": `query getContainerRepositoryTags($id: ContainerRepositoryID!, $first: Int, $last: Int, $after: String, $before: String, $name: String, $sort: ContainerRepositoryTagSort, $referrers: Boolean = false) {
-  containerRepository(id: $id) {
-    tags(
-      after: $after
-      before: $before
-      first: $first
-      last: $last
-      name: $name
-      sort: $sort
-      referrers: $referrers
-    ) {
-      nodes {
-        digest
-        location
-        path
-        name
-        createdAt
-        publishedAt
-      }
-    }
-  }
-}`})
-	if err != nil {
-		return nil, err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, "https://gitlab.com/api/graphql", bytes.NewReader(payload))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	res, err := c.Client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer res.Body.Close()
-
-	if err := httputil.AssertStatusCode(res, http.StatusOK); err != nil {
-		return nil, err
-	}
-
-	var result struct {
-		Errors []any `json:"error"`
-		Data   struct {
-			ContainerRepository struct {
-				Tags struct {
-					Nodes []ContainerRepositoryTag `json:"nodes"`
-				} `json:"tags"`
-			} `json:"containerRepository"`
-		} `json:"data"`
-	}
-	if err := json.NewDecoder(res.Body).Decode(&result); err != nil {
-		return nil, err
-	} else if len(result.Errors) > 0 {
-		return nil, fmt.Errorf("got unexpected graphql error: %v", result.Errors)
-	}
-
-	return result.Data.ContainerRepository.Tags.Nodes, nil
 }
