+++ internal/htmlutil/attr.go
@@ -2,6 +2,8 @@ package htmlutil
 
 import "golang.org/x/net/html"
 
+// GetAttr returns a node's attribute value by key.
+// Returns false if the attribute does not exist.
 func GetAttr(node *html.Node, key string) (string, bool) {
 	for _, attr := range node.Attr {
 		if attr.Key == key {
@@ -12,6 +14,7 @@ func GetAttr(node *html.Node, key string) (string, bool) {
 	return "", false
 }
 
+// SetAttr sets a node's attribute value by key.
 func SetAttr(node *html.Node, key string, value string) {
 	for i, attr := range node.Attr {
 		if attr.Key == key {
