+++ internal/oci/annotations.go
@@ -5,8 +5,11 @@ import (
 	"time"
 )
 
+// Annotations holds OCI annotations.
 type Annotations map[string]string
 
+// Source returns the value of the standard OCI source annotations.
+// The annotation should point to a source code repository.
 func (a Annotations) Source() string {
 	if a == nil {
 		return ""
@@ -20,6 +23,9 @@ func (a Annotations) Source() string {
 	return s
 }
 
+// CreatedTime returns value of the standard OCI created time annotations.
+// The annotation should hold the time at which the image or artifact was
+// created.
 func (a Annotations) CreatedTime() time.Time {
 	s := a["org.opencontainers.image.created"]
 	if s == "" {
@@ -32,16 +38,22 @@ func (a Annotations) CreatedTime() time.Time {
 	return time
 }
 
+// URL returns value of the standard OCI URL annotation.
+// The annotation should hold a URL to the project's website.
 func (a Annotations) URL() string {
 	return a["org.opencontainers.image.url"]
 }
 
+// DocumentationURL returns value of the standard OCI documentation URL
+// annotation.
+// The annotation should hold a URL to the project's documentation.
 func (a Annotations) DocumentationURL() string {
 	return a["org.opencontainers.image.documentation"]
 }
 
 // Merge returns the merge of a and b.
 // If both are nil, nil is returned.
+// If the values exist in both sets, values in b takes precedence.
 func (a Annotations) Merge(b Annotations) Annotations {
 	if a == nil && b == nil {
 		return nil
