+++ internal/otelutil/const.go
@@ -1,5 +1,6 @@
 package otelutil
 
 const (
+	// DefaultScope is the scope used by Cupdate by default.
 	DefaultScope string = "github.com/AlexGustafsson/cupdate"
 )
