+++ internal/quay/client_test.go
@@ -17,14 +17,14 @@ func TestGetScan(t *testing.T) {
 		t.Skip()
 	}
 
-	ref, err := oci.ParseReference("quay.io/openshift-release-dev/ocp-release")
+	ref, err := oci.ParseReference("quay.io/openshift-release-dev/ocp-release@sha256:7708f832ae02919f2cdb2798fdbc64e17ce7a576d1e3baabdd78a000d2d62f40")
 	require.NoError(t, err)
 
 	client := &Client{
 		Client: httputil.NewClient(cachetest.NewCache(t), 24*time.Hour),
 	}
 
-	scan, err := client.GetScan(context.TODO(), ref, "sha256:7708f832ae02919f2cdb2798fdbc64e17ce7a576d1e3baabdd78a000d2d62f40")
+	scan, err := client.GetScan(context.TODO(), ref)
 	require.NoError(t, err)
 
 	fmt.Printf("%+v\n", scan)
