+++ internal/worker/queue.go
@@ -9,6 +9,12 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 )
 
+var _ prometheus.Collector = (*Queue[any])(nil)
+
+// Queue is a queue data structure implementation useful for tracking jobs.
+// It enables multiple producers to queue items for multiple consumers.
+// Consumers will sleep whilst they're waiting for jobs.
+// The implementation also provides token bucket-based rate limiting.
 type Queue[T comparable] struct {
 	cond    *sync.Cond
 	backlog []T
@@ -19,6 +25,12 @@ type Queue[T comparable] struct {
 	lengthGauge prometheus.Gauge
 }
 
+// NewQueue returns a new [Queue].
+// An initialized queue should be closed by calling [Queue.Close]
+//
+//   - burst controls the target number of items that can be taken from the
+//     queue.
+//   - tick is the target mean time between items being taken from the queue.
 func NewQueue[T comparable](burst int, tick time.Duration) *Queue[T] {
 	q := &Queue[T]{
 		cond:    sync.NewCond(&sync.Mutex{}),
@@ -60,6 +72,8 @@ func NewQueue[T comparable](burst int, tick time.Duration) *Queue[T] {
 	return q
 }
 
+// Push puts one or more items at the end of the queue.
+// Panics if the queue is closed.
 func (q *Queue[T]) Push(items ...T) {
 	q.cond.L.Lock()
 	defer q.cond.L.Unlock()
@@ -79,6 +93,10 @@ func (q *Queue[T]) Push(items ...T) {
 	q.cond.Broadcast()
 }
 
+// Pull returns an iterator which will feed a consumer with items.
+// The iterator will wait for new items if the queue is emptied and is only
+// closed when the queue itself is closed, or when the consumer decides to stop
+// processing items.
 func (q *Queue[T]) Pull() iter.Seq[T] {
 	return func(yield func(T) bool) {
 		q.cond.L.Lock()
@@ -105,14 +123,20 @@ func (q *Queue[T]) Pull() iter.Seq[T] {
 	}
 }
 
+// Len returns the size of the queue.
 func (q *Queue[T]) Len() int {
 	return len(q.backlog)
 }
 
+// AvailableBurst returns the number of items that could currently be taken from
+// the queue without being rate limited.
 func (q *Queue[T]) AvailableBurst() int {
 	return q.tokens
 }
 
+// Close closes the queue and frees resources.
+// Closing a queue will wake any waiting consumer and exit their pull loops.
+// Continuing to use a closed queue will panic.
 func (q *Queue[T]) Close() {
 	q.cond.L.Lock()
 	defer q.cond.L.Unlock()
@@ -121,13 +145,13 @@ func (q *Queue[T]) Close() {
 	q.cond.Broadcast()
 }
 
-// Collect implements [prometheus.Collector].
+// Collect implements prometheus.Collector.
 func (w *Queue[T]) Collect(ch chan<- prometheus.Metric) {
 	w.burstGauge.Collect(ch)
 	w.lengthGauge.Collect(ch)
 }
 
-// Describe implements [prometheus.Collector].
+// Describe implements prometheus.Collector.
 func (w *Queue[T]) Describe(descs chan<- *prometheus.Desc) {
 	prometheus.DescribeByCollect(w, descs)
 }
