+++ internal/workflow/package.go
@@ -1,0 +1,3 @@
+// Package workflow provides a generic way to represent running tasks with or
+// without dependencies. The implementation mimics GitHub actions / workflows.
+package workflow
