+++ web/hooks/useMountedDuration.ts
@@ -1,5 +1,9 @@
 import { useEffect, useRef, useState } from 'react'
 
+/**
+ * useMountedDuration returns then number of seconds the component has been
+ * mounted. Useful to show loaders if loading takes longer than expected.
+ */
 export function useMountedDuration(): number {
   const startedRef = useRef(performance.now())
   const [duration, setDuration] = useState(0)
